+++ src/pyvutils/string_utils/splitters.py
@@ -1,0 +1,11 @@
+"""
+Input Splitter Functions
+"""
+
+
+def split_by_char(string: str, char: str) -> list:
+    return [part.strip() for part in string.split(char)]
+
+
+def split_by_comma(string: str) -> list:
+    return split_by_char(string, char=",")