+++ HISTORY.md
@@ -2,11 +2,22 @@
 
 <br>
 
+## 0.1.1 (2024/03/02)
+
+- Various New Utilities
+- String Utils >> Splitters
+    - New: `split_by_char()` & `split_by_comma()` Functions
+- Parser Utils >> Numeric Parsers
+    - New: `parse_numeric_value()` & `parse_integer_value()` & `parse_float_value()`
+- List Utils >> Processors
+    - New: `remove_duplicates()`
+
+
 ## 0.1.0 (2024/02/29)
 
 - Initial Running Version
 - Set Core Design & Guidelines
     - Based Import Structure
 - Available for Limited Use & Initial Tests
-    - `assert_input()` Method from **assert_utils**
+    - `assert_input()` Function from **assert_utils**